Finished fixing displays

+++ client/src/components/RatingCard.vue
@@ -7,12 +7,13 @@
           {{rateData.point}}
       </div>
       <div>
-          <button @click.prevent="deleteRate(rateData.id)">delete</button>
+          <button class="deleteButton" @click.prevent="deleteRate(rateData.id)"><i class="far fa-trash-alt"></i></button>
       </div>
   </div>
 </template>
 
 <script>
+import Swal from 'sweetalert2'
 export default {
   name: 'RatingCard',
   props: {
@@ -22,16 +23,39 @@ export default {
   },
   methods: {
       deleteRate (id) {
-          this.$store.dispatch('deleteRate', {
-            id: id,
-            movieId: this.$route.params.id
-          })
+          Swal.fire({
+            title: 'Are you sure you want to delete this Rate?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+            }).then((result) => {
+            if (result.value) {
+                this.$store.dispatch('deleteRate', {
+                    id: id,
+                    movieId: this.$route.params.id
+                })
+            }
+            })
       }
   }
 }
 </script>
 
 <style scoped>
+.deleteButton {
+    border: none;
+    border-radius: 10px;
+    background-color: #fa6464;
+    cursor: pointer;
+    width: 4vw;
+    height: 4vh;
+}
+.deleteButton:hover {
+    background-color: #c75050;
+}
 .rateCard {
     height: 5vh;
     width: 60vh;
